requirements.txt
Django==1.11.4
django-registration-redux==1.7
psycopg2==2.7.3.1
pytz==2017.2
